+++ src/main/java/org/md/api/s3/config/SwaggerConfiguration.java
@@ -13,6 +13,9 @@ import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
+/**
+ * set up the swagger docs basic overview
+ */
 @Configuration
 @EnableSwagger2
 public class SwaggerConfiguration {