+++ src/main/java/org/md/api/s3/model/exception/AwsS3GeneralException.java
@@ -1,0 +1,14 @@
+package org.md.api.s3.model.exception;
+
+public class AwsS3GeneralException extends Exception {
+
+    private static final long serialVersionUID = 1L;
+
+    public AwsS3GeneralException() {
+        super();
+    }
+    
+    public AwsS3GeneralException(String message) {
+        super(message);
+    }
+}