+++ src/main/java/org/md/api/s3/model/exception/MissingAwsInformation.java
@@ -1,6 +1,6 @@
 package org.md.api.s3.model.exception;
 
-public class MissingAwsInformation extends Exception {
+public class MissingAwsInformation extends AwsS3GeneralException {
 
 	private static final long serialVersionUID = 1L;