+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved validations to own classes
 - Class package naming convention
 
+### Removed
+- Removed specific object (index files) within S3 being modified
+
 ## [0.0.1-SNAPSHOT] - 2020-01-26
 ### Added
 - Initial project creation